requirements.txt
streamlit
yfinance
pandas
plotly
numpy
python-dotenv
requests